separate writer to another package namespace
