+++ torch/utils/tensorboardX/__init__.py
@@ -1,6 +1,6 @@
 """A module for visualization with tensorboard
 """
 
-from .record_writer import RecordWriter
+from ..tensorboardnotf.record_writer import RecordWriter
 from .torchvis import TorchVis
 from .writer import FileWriter, SummaryWriter
